songs_controller.rb: get_zip will not download more than 500 zip files in one month

+++ app/controllers/songs_controller.rb
@@ -34,40 +34,30 @@ class SongsController < ApplicationController
   end
 
   def get_zip
-    access_key_id = Rails.application.credentials.development[:aws][:access_key_id]
-    secret_access_key = Rails.application.credentials.development[:aws][:secret_access_key]
-    s3 = Aws::S3::Resource.new(region: 'us-east-1', access_key_id: access_key_id, secret_access_key: secret_access_key)
-    filename = Song.find(params[:id]).filename << ".zip"
-    s3_file_path ="m4a/#{filename}"
-    object = s3.bucket('cantandoinglesbucket').object(s3_file_path)
-    object.get(response_target: "#{Rails.root}/app/assets/songs/#{filename}")
-    File.chmod(0666, "#{Rails.root}/app/assets/songs/#{filename}")
-    send_file "app/assets/songs/#{filename}", :filename => "#{filename}", :url_based_filename => false, :type=>"application/zip"
-    
-    song = Song.find(params[:id])
-    downloadCount = DownloadCount.find_by(song_id: 19, month: Date.current.beginning_of_month)
-    if (downloadCount)
-      downloadCount.month_total = downloadCount.month_total + 1
-      downloadCount.save
-    else
-      newDownloadCount = DownloadCount.new(:song_id => song.id, month: Date.current.beginning_of_month, :month_total => 1)
-      newDownloadCount.save
+    if (DownloadCount.where(:month => Date.current.beginning_of_month).sum(:month_total) > 500)
+      flash[:notice] = "Already have 500 downloads"
+      redirect_to(songs_path)  
+    else 
+      access_key_id = Rails.application.credentials.development[:aws][:access_key_id]
+      secret_access_key = Rails.application.credentials.development[:aws][:secret_access_key]
+      s3 = Aws::S3::Resource.new(region: 'us-east-1', access_key_id: access_key_id, secret_access_key: secret_access_key)
+      filename = Song.find(params[:id]).filename << ".zip"
+      s3_file_path ="m4a/#{filename}"
+      object = s3.bucket('cantandoinglesbucket').object(s3_file_path)
+      object.get(response_target: "#{Rails.root}/app/assets/songs/#{filename}")
+      File.chmod(0666, "#{Rails.root}/app/assets/songs/#{filename}")
+      send_file "app/assets/songs/#{filename}", :filename => "#{filename}", :url_based_filename => false, :type=>"application/zip"
+      song = Song.find(params[:id])
+      downloadCount = DownloadCount.find_by(song_id: 19, month: Date.current.beginning_of_month)
+      if (downloadCount)
+        downloadCount.month_total = downloadCount.month_total + 1
+        downloadCount.save
+      else
+        newDownloadCount = DownloadCount.new(:song_id => song.id, month: Date.current.beginning_of_month, :month_total => 1)
+        newDownloadCount.save
+      end
+      File.delete('app/assets/songs/${filename}') if File.exists?('app/assets/songs/#{filename}')
     end
-    File.delete('app/assets/songs/${filename}') if File.exists?('app/assets/songs/#{filename}')
   end
 
-  def dealWithDownloads
-   @today = Date.today
-    @beginning_of_month = @today.beginning_of_month
-    @this_months_downloads = Download.where(:created_at => @beginning_of_month.beginning_of_day..@today.end_of_day)
-    if (@this_months_downloads.length>3999 && this.months_downloads.length<4001)
-      Notifier.num_downloads(@this_months_downloads.length).deliver
-    end
-    if (@this_months_downloads.length>4999 && this.months_downloads.length<5001)
-      Notifier.num_downloads(@this_months_downloads.length).deliver
-    end
-    if (@this_months_downloads.length>4999)
-      flash[:notice] = "Already have 5000 downloads"
-    end
-  end
 end
